Explicit float32 to int color conv; match all Go and V (except A12-1)

+++ rtiow/V/A03.v
@@ -13,9 +13,9 @@ fn main() {
 	for j := ny-1; j >= 0; j-- {
 		for i := 0; i < nx; i++ {
 			col := vec.Vec3{f32(i) / f32(nx), f32(j) / f32(ny), 0.2}
-			ir := int(255.99 * col.x)
-			ig := int(255.99 * col.y)
-			ib := int(255.99 * col.z)
+			ir := int(f32(255.99) * col.x)
+			ig := int(f32(255.99) * col.y)
+			ib := int(f32(255.99) * col.z)
 			println('$ir $ig $ib')
 		}
 	}
